# Conflicts:
#   src/main/resources/application.yaml
#   src/main/resources/templates/Symptom/symptom-history.html

+++ src/main/java/withus/entity/Tbl_weight.java
@@ -19,7 +19,7 @@ public class Tbl_weight {
     private RecordKey pk;
 
     @Column(name = "weight")
-    private float weight;
+    private Integer weight;
 
     @Column(name = "week")
     private Integer week;
